# app/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an api api for Paperless.dev",
        "title": "Paperless.dev API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/hwangseonu/paperless.dev/refs/heads/main/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/resumes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get all resumes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "get all resumes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner ID of resumes",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "resumes": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/schema.ResumeResponseSchema"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "create new resume",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "create new resume",
                "parameters": [
                    {
                        "description": "initial values of resume",
                        "name": "resume",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.ResumeCreateSchema"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "resume": {
                                    "$ref": "#/definitions/schema.ResumeResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    }
                }
            }
        },
        "/resumes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get resume by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "get resume by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resume ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "resume": {
                                    "$ref": "#/definitions/schema.ResumeResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete resume by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "delete resume by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resume ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update resume by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "update resume by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resume ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update values of resume",
                        "name": "resume",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.ResumeUpdateSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "resume": {
                                    "$ref": "#/definitions/schema.ResumeResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    }
                }
            }
        },
...
        "/users/register": {
            "post": {
                "description": "Starts the sign-up process by creating a temporary user profile in the cache and dispatching a verification code to the provided email address.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Initiate user registration \u0026 send verification email",
                "parameters": [
                    {
                        "description": "initial values of User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UserCreateSchema"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schema.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.EducationResponseSchema": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "string"
                },
                "degree": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "gpa": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "school": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
        "schema.EducationUpdateSchema": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "string"
                },
                "degree": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "gpa": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "school": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
        "schema.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "schema.ExperienceResponseSchema": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "schema.ExperienceUpdateSchema": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "schema.ProjectResponseSchema": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "startDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "schema.ProjectUpdateSchema": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "startDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "schema.ResumeCreateSchema": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "template": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "schema.ResumeResponseSchema": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "educations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.EducationResponseSchema"
                    }
                },
                "email": {
                    "type": "string"
                },
                "experiences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ExperienceResponseSchema"
                    }
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ProjectResponseSchema"
                    }
                },
                "public": {
                    "type": "boolean"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "template": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "schema.ResumeUpdateSchema": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "educations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.EducationUpdateSchema"
                    }
                },
                "email": {
                    "type": "string"
                },
                "experiences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ExperienceUpdateSchema"
                    }
                },
                "image": {
                    "type": "string"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ProjectUpdateSchema"
                    }
                },
                "public": {
                    "type": "boolean"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "template": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "schema.UserCreateSchema": {
            "type": "object",
            "required": [
                "email",
                "nickname",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Type 'Bearer ' followed by your API key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}